+++ src/utils/archer_atof.c
@@ -6,7 +6,7 @@
 /*   By: maandria <maandria@student.42antananari    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/09 08:38:47 by atolojan          #+#    #+#             */
-/*   Updated: 2025/01/21 15:22:18 by maandria         ###   ########.fr       */
+/*   Updated: 2025/03/18 14:25:31 by maandria         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -31,14 +31,14 @@ static void	init_atof(int *i, int *j, int *sign, int *comma)
 	*comma = 0;
 }
 
-static void	set_add(float *nbr, float to_add, int i[])
+static void	set_add(double *nbr, double to_add, int i[])
 {
 	*nbr += to_add;
 	i[1]++;
 	i[0]++;
 }
 
-static void	atof_tools(const char *nptr, int i[], float *nbr)
+static void	atof_tools(const char *nptr, int i[], double *nbr)
 {
 	while ((nptr[i[0]] >= '0' && nptr[i[0]] <= '9') || nptr[i[0]] == '.')
 	{
@@ -65,10 +65,10 @@ static void	atof_tools(const char *nptr, int i[], float *nbr)
 	}
 }
 
-float	arc_atof(const char *nptr)
+double	arc_atof(const char *nptr)
 {
 	int		i[4];
-	float	nbr;
+	double	nbr;
 
 	nbr = 0;
 	init_atof(&i[0], &i[1], &i[2], &i[3]);
